# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Apartment Hunting Tracker API",
    "description": "API for tracking apartment listings and scheduled viewings, secured with GitHub OAuth. Log in via GET /auth/github before trying protected (lock icon) routes below.",
    "version": "1.0.0"
  },
  "host": "apartment-tracker-api-4yal.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "cookieAuth": {
      "type": "apiKey",
      "in": "cookie",
      "name": "connect.sid",
      "description": "Visit GET /auth/github in this same browser tab first and approve on GitHub. The session cookie that sets is then sent automatically with requests made from this page."
    }
  },
  "paths": {
...
    "/auth/success": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}